Updating the incident updates

+++ plugins/harassmap/contact/updates/update_small_contact_form_messages.php
@@ -9,21 +9,21 @@ class UpdateSmallContactFormMessages extends Migration
 {
     public function up()
     {
-        Schema::table('janvince_smallcontactform_messages', function ($table) {
-
-            $table->integer('domain_id')->unsigned();
-            $table->foreign('domain_id')->references('id')->on('harassmap_incidents_domain');
-
-        });
+        if (Schema::hasTable('janvince_smallcontactform_messages')) {
+            Schema::table('janvince_smallcontactform_messages', function ($table) {
+                $table->integer('domain_id')->unsigned();
+                $table->foreign('domain_id')->references('id')->on('harassmap_incidents_domain');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('janvince_smallcontactform_messages', function ($table) {
-
-            $table->dropForeign('janvince_smallcontactform_messages_domain_id_foreign');
-            $table->dropColumn('domain_id');
-
-        });
+        if (Schema::hasTable('janvince_smallcontactform_messages')) {
+            Schema::table('janvince_smallcontactform_messages', function ($table) {
+                $table->dropForeign('janvince_smallcontactform_messages_domain_id_foreign');
+                $table->dropColumn('domain_id');
+            });
+        }
     }
 }
